extract site header as a sole component
