+++ confiture-web-app/src/App.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import { ref, watch, computed } from "vue";
-import { useRoute, RouteLocationRaw } from "vue-router";
+import { ref, watch } from "vue";
+import { useRoute } from "vue-router";
 import { useHead } from "@vueuse/head";
 
 import Breadcrumb, { BreadcrumbLink } from "./components/Breadcrumb.vue";
 import ToastNotification from "./components/ToastNotification.vue";
-import { useAuditStore, useReportStore } from "./store";
+import SiteHeader from "./components/SiteHeader.vue";
 
 const route = useRoute();
 
@@ -43,57 +43,6 @@ useHead({
     { name: "twitter:card", content: "summary_large_image" },
   ],
 });
-
-const reportStore = useReportStore();
-const auditStore = useAuditStore();
-
-const homeLocation = { label: "Accueil", to: { name: "home" } };
-const helpLocation = { label: "Aide", to: { name: "help" } };
-const resourcesLocation = { label: "Ressources", to: { name: "resources" } };
-
-const menuItems = computed<Array<{ to: RouteLocationRaw; label: string }>>(
-  () => {
-    if (reportStore.data) {
-      const reportLocation = {
-        to: {
-          name: "report",
-          params: { uniqueId: reportStore.data.consultUniqueId },
-        },
-        label: "Rapport d’audit",
-      };
-      return [reportLocation, resourcesLocation, helpLocation];
-    }
-
-    if (auditStore.data) {
-      const auditLocation = {
-        label: `Audit ${auditStore.data.procedureName}`,
-        to: auditStore.lastVisitedStepLocation ?? {
-          name: "edit-audit-step-one",
-          params: { uniqueId: auditStore.data.editUniqueId },
-        },
-      };
-      return [homeLocation, auditLocation, resourcesLocation];
-    }
-
-    return [homeLocation, resourcesLocation, helpLocation];
-  }
-);
-
-const logoLink = computed(() => {
-  if (reportStore.data) {
-    return {
-      route: {
-        name: "report",
-        params: { uniqueId: reportStore.data.consultUniqueId },
-      },
-      title: "Rapport d’audit - Ara",
-    };
-  }
-  return {
-    route: { name: "home" },
-    title: "Accueil - Ara",
-  };
-});
 </script>
 
 <template>
@@ -110,86 +59,20 @@ const logoLink = computed(() => {
     </nav>
   </div>
 
-  <header id="header" role="banner" class="fr-header">
-    <div class="fr-header__body">
-      <div class="fr-container">
-        <div class="fr-header__body-row">
-          <div class="fr-header__brand fr-enlarge-link">
-            <div class="fr-header__brand-top">
-              <div class="fr-header__logo">
-                <p class="fr-logo">
-                  République
-                  <br />Française
-                </p>
-              </div>
-              <div class="fr-header__navbar">
-                <button
-                  id="fr-btn-menu-mobile"
-                  class="fr-btn--menu fr-btn"
-                  data-fr-opened="false"
-                  aria-controls="modal-navigation-header"
-                  aria-haspopup="menu"
-                  title="Menu"
-                >
-                  Menu
-                </button>
-              </div>
-            </div>
-            <div class="fr-header__service">
-              <RouterLink :to="logoLink.route" :title="logoLink.title">
-                <p class="fr-header__service-title">
-                  ARA
-                  <span
-                    class="fr-badge fr-badge--sm fr-badge--info fr-badge--no-icon"
-                    >BÊTA</span
-                  >
-                </p>
-              </RouterLink>
-              <p class="fr-header__service-tagline">
-                Réaliser des audits d’accessibilité
-              </p>
-            </div>
-          </div>
-        </div>
-      </div>
+  <SiteHeader />
 
-      <div
-        id="modal-navigation-header"
-        class="fr-header__menu fr-modal"
-        aria-labelledby="fr-btn-menu-mobile"
-      >
-        <div class="fr-container">
-          <button
-            class="fr-link--close fr-link"
-            aria-controls="modal-navigation-header"
-          >
-            Fermer
-          </button>
-          <div class="fr-header__menu-links"></div>
-          <nav class="fr-nav" role="navigation" aria-label="Menu principal">
-            <ul class="fr-nav__list">
-              <li
-                v-for="item in menuItems"
-                :key="item.label"
-                class="fr-nav__item"
-              >
-                <RouterLink class="fr-nav__link" :to="item.to">
-                  {{ item.label }}
-                </RouterLink>
-              </li>
-            </ul>
-          </nav>
-        </div>
-      </div>
-    </div>
-  </header>
   <div class="fr-notice fr-notice--info">
     <div class="fr-container">
       <div class="fr-notice__body">
         <p class="fr-notice__title">
           Vos avis nous sont précieux pour améliorer cet outil, n’hésitez pas à
           nous faire part de vos retours depuis ce
-          <RouterLink :to="{ name: 'feedback' }" title="court formulaire - nouvelle fenêtre" target="_blank">court formulaire</RouterLink>
+          <RouterLink
+            :to="{ name: 'feedback' }"
+            title="court formulaire - nouvelle fenêtre"
+            target="_blank"
+            >court formulaire</RouterLink
+          >
         </p>
       </div>
     </div>
@@ -217,10 +100,18 @@ const logoLink = computed(() => {
         <div class="fr-footer__content">
           <p class="fr-footer__content-desc">
             Ce site est réalisé par
-            <a href="https://design.numerique.gouv.fr/" title="DesignGouv - nouvelle fenêtre" target="_blank" rel="noreferrer noopener"
+            <a
+              href="https://design.numerique.gouv.fr/"
+              title="DesignGouv - nouvelle fenêtre"
+              target="_blank"
+              rel="noreferrer noopener"
               >DesignGouv</a
             >, le pôle Design des services numériques de la
-            <a href="https://www.numerique.gouv.fr/dinum/" title="direction interministérielle du numérique - nouvelle fenêtre"  rel="noreferrer noopener"  target="_blank"
+            <a
+              href="https://www.numerique.gouv.fr/dinum/"
+              title="direction interministérielle du numérique - nouvelle fenêtre"
+              rel="noreferrer noopener"
+              target="_blank"
               >direction interministérielle du numérique</a
             >.
           </p>
@@ -313,7 +204,8 @@ const logoLink = computed(() => {
             Sauf mention contraire, tous les contenus de ce site sont sous
             <a
               href="https://github.com/etalab/licence-ouverte/blob/master/LO.md"
-              target="_blank"  rel="noreferrer noopener"
+              target="_blank"
+              rel="noreferrer noopener"
               title="licence etalab-2.0 - nouvelle fenêtre"
               >licence etalab-2.0</a
             >
